# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pusk API
  version: 0.7.0
  description: |
    Self-hosted alert platform for ops teams. Telegram Bot API compatible.

    ## Authentication

    Three auth schemes are used depending on the endpoint group:

    - **Bot Token** — passed as a path parameter in `/bot/{token}/{method}` endpoints.
    - **JWT Bearer** — returned by `/api/auth` or `/api/register`, used for all `/api/` client endpoints.
    - **Admin Token** — server-configured token for `/admin/` management endpoints. Some admin endpoints also accept JWT with admin role.
  license:
    name: BSL 1.1
    url: https://github.com/getpusk/pusk/blob/main/LICENSE
  contact:
    name: Pusk
    url: https://getpusk.ru
servers:
  - url: https://getpusk.ru
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth or /api/register. 7-day TTL.
    AdminToken:
      type: apiKey
      in: header
      name: Authorization
      description: Admin token configured on the server. Passed as `Bearer {admin_token}`.
  schemas:
...
    InviteResponse:
      type: object
      properties:
        code:
          type: string
        url:
          type: string
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
...
paths:
...
  /api/invite:
    post:
      tags: [Client Infra]
      operationId: clientInviteCreate
      summary: Create an invite link (admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Invite created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '403':
          description: Not an admin.
    delete:
      tags: [Client Infra]
      operationId: clientInviteRevoke
      summary: Revoke an invite code (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Invite revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '403':
          description: Not an admin.
...